answer vote bug fixed

+++ AskMate/AskMate/Controllers/HomeController.cs
@@ -159,14 +159,14 @@ namespace AskMate.Controllers
             return View("Question", question);
         }
 
-        public IActionResult A_UpVote([FromRoute(Name = "id")]int id)
+        public IActionResult A_UpVote([FromRoute(Name = "id")]long id)
         {
             AnswerModel answer = Idao.GetAnswerByUnique(id);
             answer.VoteUp();
             return View("Question", Idao.GetQuestionById(answer.Question_Id));
         }
 
-        public IActionResult A_DownVote([FromRoute(Name = "id")]int id)
+        public IActionResult A_DownVote([FromRoute(Name = "id")]long id)
         {
             AnswerModel answer = Idao.GetAnswerByUnique(id);
             answer.VoteDown();
